requirements.txt
Flask==2.2.5
Flask-Limiter==3.5.0
requests==2.31.0
Pillow==9.5.0
python-dotenv==1.0.1
Flask-Cors